week6 do 14-12-2017 is_numeric

+++ delete.php
@@ -1,13 +1,15 @@
 <?php
-  // Maak contact met de server en database;
-  include("./db_connect.php");
-
-  // De sql delete query
-  $sql = "DELETE FROM `users` WHERE `id` = " . $_GET["id"];
-
-  // Vuur de query af op de database
-  $result = mysqli_query($conn, $sql);
-
+  if (is_numeric($_GET["id"])) {
+    // Maak contact met de server en database;
+    include("./db_connect.php");
+  
+    // De sql delete query
+    $sql = "DELETE FROM `users` WHERE `id` = " . $_GET["id"];
+  
+    // Vuur de query af op de database
+    $result = mysqli_query($conn, $sql);
+  }
+  
   // Leid om naar index.php
   header("Location: ./index.php");
 ?>
